django_authenticator_keycloak/requirements.txt
django
djangorestframework
requests
oic

